+++ srcs/ft_memcmp.c
@@ -5,7 +5,7 @@ int		ft_memcmp(void *buf1, void *buf2, size_t count)
 	int i;
 
 	i = 0;
-	while (i != (count - 1))
+	while (i != (int)(count - 1))
 	{
 		if ((((char*)buf1)[i]) && (((char*)buf1)[i]))
 			if (((char*)buf1)[i] != ((char*)buf2)[i])
