+++ srcs/ft_strncat.c
@@ -9,7 +9,7 @@ char	*ft_strncat(char *str1, char *str2, size_t count)
 	j = -1;
 	while (str1[i])
 		i++;
-	while(str2[++j] && j < count)
+	while(str2[++j] && j < (int)count)
 		str1[i++] = str2[j];
 	return (str1);
 }
