+++ srcs/ft_strncpy.c
@@ -5,7 +5,7 @@ char 	*ft_strncpy( char * destptr, const char * srcptr, size_t num )
 	int i;
 
 	i = -1;
-	while (srcptr[++i] && i < num)
+	while (srcptr[++i] && i < (int)num)
 		destptr[i] = srcptr[i];
 	if (srcptr[i] == '\0')
 		(destptr[i] = '\0');
