+++ srcs/ft_strnequ.c
@@ -5,12 +5,12 @@ int 	ft_strnequ(char const *s1, char const *s2, size_t n)
 	int i;
 
 	i = -1;
-	while (s1[++i] && s2[i] && i < n)
+	while (s1[++i] && s2[i] && i < (int)n)
 	{
 		if (s1[i] != s2[i])
 			return (0);
 	}
-	if (i == n - 1)
+	if (i == (int)(n - 1))
 		if (s1[i] == s2[i])
 			return (1);
 	return (0);
